file upload

+++ app/Http/Controllers/PostsController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Storage;
 use App\Post;
 use DB;
 
@@ -17,6 +18,7 @@ class PostsController extends Controller
     {
         $this->middleware('auth', ['except' => ['index', 'show']]);
     }
+
     /**
      * Display a listing of the resource.
      *
@@ -24,9 +26,8 @@ class PostsController extends Controller
      */
     public function index()
     {
-        // $posts = Post::all();
-        // $posts = DB::select('SELECT * FROM posts');
-        $posts = Post::orderby('id', 'desc')->get();
+        // $posts = Post::orderBy('title','asc')->get();
+        $posts = DB::select('SELECT * FROM posts');
         return view('posts.index')->with('posts', $posts);
     }
 
@@ -48,20 +49,37 @@ class PostsController extends Controller
      */
     public function store(Request $request)
     {
-        // $post = Post::create($request->validate([
-        //     'title' => 'required',
-        //     'body' => 'required'
-        // ]));
-        $this->validate($request, [
+        $this->validate($request,[
             'title' => 'required',
-            'body' => 'required'
+            'body' => 'required',
+            'cover_image' => 'image|nullable|max:1999'
         ]);
-        $post = new Post();
-
+        // handle the file upload
+        if($request->hasFile('cover_image'))
+        {
+            // get filename with extension
+            $filenameWithExt = $request->file('cover_image')->getClientOriginalName();
+            // get just filename
+            $filename = pathinfo($filenameWithExt, PATHINFO_FILENAME);
+            // get just ext
+            $extension = $request->file('cover_image')->getClientOriginalExtension();
+            // filename to store
+            $fileNameToStore = $filename.'_'.time().'.'.$extension;
+            // upload image
+            $path = $request->file('cover_image')->storeAs('public/cover_images',$fileNameToStore);
+        } else
+        {
+            $fileNameToStore = 'noimage.jpg';
+        }
+        
+        
+        $post = new Post;
         $post->title = $request->input('title');
         $post->body = $request->input('body');
         $post->user_id = auth()->user()->id;
+        $post->cover_image = $fileNameToStore;
         $post->save();
+
         return redirect('/posts')->with('success', 'Post created.');
     }
 
@@ -71,9 +89,10 @@ class PostsController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function show(Post $post)
+    public function show($id)
     {
-        return view('posts.show')->with('post', $post);
+        $post = Post::find($id);
+        return view('posts.show')->with('post',$post);
     }
 
     /**
@@ -85,11 +104,10 @@ class PostsController extends Controller
     public function edit($id)
     {
         $post = Post::find($id);
-        if (auth()->user()->id !== $post->user_id) {
-            return redirect('/posts')->with('error', "Unauthorized page.");
+        if(auth()->user()->id !== $post->user_id) {
+            return redirect('/posts')->with('error', 'Unauthrized page.');
         }
-
-        return view('posts.edit')->with('post', $post);
+        return view('posts.edit')->with('post',$post);
     }
 
     /**
@@ -101,14 +119,32 @@ class PostsController extends Controller
      */
     public function update(Request $request, $id)
     {
-        $this->validate($request, [
+        $this->validate($request,[
             'title' => 'required',
-            'body' => 'required'
+            'body' => 'required',
+            'cover_image' => 'image|nullable|max:1999'
         ]);
-        $post = Post::find($id);
+        // handle the file upload
+        if($request->hasFile('cover_image'))
+        {
+            // get filename with extension
+            $filenameWithExt = $request->file('cover_image')->getClientOriginalName();
+            // get just filename
+            $filename = pathinfo($filenameWithExt, PATHINFO_FILENAME);
+            // get just ext
+            $extension = $request->file('cover_image')->getClientOriginalExtension();
+            // filename to store
+            $fileNameToStore = $filename.'_'.time().'.'.$extension;
+            // upload image
+            $path = $request->file('cover_image')->storeAs('public/cover_images',$fileNameToStore);
+        }
 
+        $post = Post::find($id);
         $post->title = $request->input('title');
         $post->body = $request->input('body');
+        if($request->hasFile('cover_image')) {
+            $post->cover_image = $fileNameToStore;
+        }
         $post->save();
 
         return redirect('/posts')->with('success', 'Post updated.');
@@ -123,8 +159,16 @@ class PostsController extends Controller
     public function destroy($id)
     {
         $post = Post::find($id);
-        $post->delete();
+        if(auth()->user()->id !== $post->user_id) {
+            return redirect('/posts')->with('error', 'Unauthrized page.');
+        }
 
+        if($post->cover_image != 'noimage.jpg')
+        {
+            Storage::delete('public/cover_images/'.$post->cover_image);
+        }
+        
+        $post->delete();
         return redirect('/posts')->with('success', 'Post removed.');
     }
 }
